EXUP_Removing_unwanted_components
