+++ package.json
@@ -32,6 +32,7 @@
     "react": "18.2.0",
     "react-apexcharts": "1.4.1",
     "react-dom": "18.2.0",
+    "react-excel-renderer": "^1.1.0",
     "simplebar-react": "^3.2.4",
     "yup": "1.3.2"
   },
