+++ src/layouts/dashboard/layout.js
@@ -3,7 +3,6 @@ import { usePathname } from 'next/navigation';
 import { styled } from '@mui/material/styles';
 import { withAuthGuard } from 'src/hocs/with-auth-guard';
 import { SideNav } from './side-nav';
-import { TopNav } from './top-nav';
 
 const SIDE_NAV_WIDTH = 280;
 
@@ -47,7 +46,6 @@ export const Layout = withAuthGuard((props) => {
 
   return (
     <>
-      <TopNav onNavOpen={() => setOpenNav(true)} />
       <SideNav
         onClose={() => setOpenNav(false)}
         open={openNav}
