+++ src/pages/index.js
@@ -1,227 +1,263 @@
+import { useCallback, useMemo, useState } from 'react';
 import Head from 'next/head';
 import { subDays, subHours } from 'date-fns';
-import { Box, Container, Unstable_Grid2 as Grid } from '@mui/material';
+import ArrowDownOnSquareIcon from '@heroicons/react/24/solid/ArrowDownOnSquareIcon';
+import ArrowUpOnSquareIcon from '@heroicons/react/24/solid/ArrowUpOnSquareIcon';
+import PlusIcon from '@heroicons/react/24/solid/PlusIcon';
+import { Box, Button, Container, Stack, SvgIcon, Typography } from '@mui/material';
+import { useSelection } from 'src/hooks/use-selection';
 import { Layout as DashboardLayout } from 'src/layouts/dashboard/layout';
-import { OverviewBudget } from 'src/sections/overview/overview-budget';
-import { OverviewLatestOrders } from 'src/sections/overview/overview-latest-orders';
-import { OverviewLatestProducts } from 'src/sections/overview/overview-latest-products';
-import { OverviewSales } from 'src/sections/overview/overview-sales';
-import { OverviewTasksProgress } from 'src/sections/overview/overview-tasks-progress';
-import { OverviewTotalCustomers } from 'src/sections/overview/overview-total-customers';
-import { OverviewTotalProfit } from 'src/sections/overview/overview-total-profit';
-import { OverviewTraffic } from 'src/sections/overview/overview-traffic';
+import { CustomersTable } from 'src/sections/customer/customers-table';
+import { CustomersSearch } from 'src/sections/customer/customers-search';
+import { applyPagination } from 'src/utils/apply-pagination';
 
 const now = new Date();
 
-const Page = () => (
-  <>
-    <Head>
-      <title>
-        Overview | Devias Kit
-      </title>
-    </Head>
-    <Box
-      component="main"
-      sx={{
-        flexGrow: 1,
-        py: 8
-      }}
-    >
-      <Container maxWidth="xl">
-        <Grid
-          container
-          spacing={3}
-        >
-          <Grid
-            xs={12}
-            sm={6}
-            lg={3}
-          >
-            <OverviewBudget
-              difference={12}
-              positive
-              sx={{ height: '100%' }}
-              value="$24k"
+const data = [
+  {
+    id: '5e887ac47eed253091be10cb',
+    address: {
+      city: 'Cleveland',
+      country: 'USA',
+      state: 'Ohio',
+      street: '2849 Fulton Street'
+    },
+    avatar: '/assets/avatars/avatar-carson-darrin.png',
+    createdAt: subDays(subHours(now, 7), 1).getTime(),
+    email: 'carson.darrin@devias.io',
+    name: 'Carson Darrin',
+    phone: '304-428-3097'
+  },
+  {
+    id: '5e887b209c28ac3dd97f6db5',
+    address: {
+      city: 'Atlanta',
+      country: 'USA',
+      state: 'Georgia',
+      street: '1865  Pleasant Hill Road'
+    },
+    avatar: '/assets/avatars/avatar-fran-perez.png',
+    createdAt: subDays(subHours(now, 1), 2).getTime(),
+    email: 'fran.perez@devias.io',
+    name: 'Fran Perez',
+    phone: '712-351-5711'
+  },
+  {
+    id: '5e887b7602bdbc4dbb234b27',
+    address: {
+      city: 'North Canton',
+      country: 'USA',
+      state: 'Ohio',
+      street: '4894  Lakeland Park Drive'
+    },
+    avatar: '/assets/avatars/avatar-jie-yan-song.png',
+    createdAt: subDays(subHours(now, 4), 2).getTime(),
+    email: 'jie.yan.song@devias.io',
+    name: 'Jie Yan Song',
+    phone: '770-635-2682'
+  },
+  {
+    id: '5e86809283e28b96d2d38537',
+    address: {
+      city: 'Madrid',
+      country: 'Spain',
+      name: 'Anika Visser',
+      street: '4158  Hedge Street'
+    },
+    avatar: '/assets/avatars/avatar-anika-visser.png',
+    createdAt: subDays(subHours(now, 11), 2).getTime(),
+    email: 'anika.visser@devias.io',
+    name: 'Anika Visser',
+    phone: '908-691-3242'
+  },
+  {
+    id: '5e86805e2bafd54f66cc95c3',
+    address: {
+      city: 'San Diego',
+      country: 'USA',
+      state: 'California',
+      street: '75247'
+    },
+    avatar: '/assets/avatars/avatar-miron-vitold.png',
+    createdAt: subDays(subHours(now, 7), 3).getTime(),
+    email: 'miron.vitold@devias.io',
+    name: 'Miron Vitold',
+    phone: '972-333-4106'
+  },
+  {
+    id: '5e887a1fbefd7938eea9c981',
+    address: {
+      city: 'Berkeley',
+      country: 'USA',
+      state: 'California',
+      street: '317 Angus Road'
+    },
+    avatar: '/assets/avatars/avatar-penjani-inyene.png',
+    createdAt: subDays(subHours(now, 5), 4).getTime(),
+    email: 'penjani.inyene@devias.io',
+    name: 'Penjani Inyene',
+    phone: '858-602-3409'
+  },
+  {
+    id: '5e887d0b3d090c1b8f162003',
+    address: {
+      city: 'Carson City',
+      country: 'USA',
+      state: 'Nevada',
+      street: '2188  Armbrester Drive'
+    },
+    avatar: '/assets/avatars/avatar-omar-darboe.png',
+    createdAt: subDays(subHours(now, 15), 4).getTime(),
+    email: 'omar.darobe@devias.io',
+    name: 'Omar Darobe',
+    phone: '415-907-2647'
+  },
+  {
+    id: '5e88792be2d4cfb4bf0971d9',
+    address: {
+      city: 'Los Angeles',
+      country: 'USA',
+      state: 'California',
+      street: '1798  Hickory Ridge Drive'
+    },
+    avatar: '/assets/avatars/avatar-siegbert-gottfried.png',
+    createdAt: subDays(subHours(now, 2), 5).getTime(),
+    email: 'siegbert.gottfried@devias.io',
+    name: 'Siegbert Gottfried',
+    phone: '702-661-1654'
+  },
+  {
+    id: '5e8877da9a65442b11551975',
+    address: {
+      city: 'Murray',
+      country: 'USA',
+      state: 'Utah',
+      street: '3934  Wildrose Lane'
+    },
+    avatar: '/assets/avatars/avatar-iulia-albu.png',
+    createdAt: subDays(subHours(now, 8), 6).getTime(),
+    email: 'iulia.albu@devias.io',
+    name: 'Iulia Albu',
+    phone: '313-812-8947'
+  },
+  {
+    id: '5e8680e60cba5019c5ca6fda',
+    address: {
+      city: 'Salt Lake City',
+      country: 'USA',
+      state: 'Utah',
+      street: '368 Lamberts Branch Road'
+    },
+    avatar: '/assets/avatars/avatar-nasimiyu-danai.png',
+    createdAt: subDays(subHours(now, 1), 9).getTime(),
+    email: 'nasimiyu.danai@devias.io',
+    name: 'Nasimiyu Danai',
+    phone: '801-301-7894'
+  }
+];
+
+const useCustomers = (page, rowsPerPage) => {
+  return useMemo(
+    () => {
+      return applyPagination(data, page, rowsPerPage);
+    },
+    [page, rowsPerPage]
+  );
+};
+
+const useCustomerIds = (customers) => {
+  return useMemo(
+    () => {
+      return customers.map((customer) => customer.id);
+    },
+    [customers]
+  );
+};
+
+const Page = () => {
+  const [page, setPage] = useState(0);
+  const [rowsPerPage, setRowsPerPage] = useState(5);
+  const customers = useCustomers(page, rowsPerPage);
+  const customersIds = useCustomerIds(customers);
+  const customersSelection = useSelection(customersIds);
+
+  const handlePageChange = useCallback(
+    (event, value) => {
+      setPage(value);
+    },
+    []
+  );
+
+  const handleRowsPerPageChange = useCallback(
+    (event) => {
+      setRowsPerPage(event.target.value);
+    },
+    []
+  );
+
+  return (
+    <>
+      <Head>
+        <title>
+          Customers | Devias Kit
+        </title>
+      </Head>
+      <Box
+        component="main"
+        sx={{
+          flexGrow: 1,
+          py: 8
+        }}
+      >
+        <Container maxWidth="xl">
+          <Stack spacing={3}>
+            <Stack
+              direction="row"
+              justifyContent="space-between"
+              spacing={4}
+            >
+              <Stack spacing={1}>
+                <Typography variant="h4">
+                  Customers
+                </Typography>
+                <Stack
+                  alignItems="center"
+                  direction="row"
+                  spacing={1}
+                >
+                  <Button
+                    color="inherit"
+                    startIcon={(
+                      <SvgIcon fontSize="small">
+                        <ArrowUpOnSquareIcon />
+                      </SvgIcon>
+                    )}
+                  >
+                    Import
+                  </Button>
+                </Stack>
+              </Stack>
+            </Stack>
+            <CustomersSearch />
+            <CustomersTable
+              count={data.length}
+              items={customers}
+              onDeselectAll={customersSelection.handleDeselectAll}
+              onDeselectOne={customersSelection.handleDeselectOne}
+              onPageChange={handlePageChange}
+              onRowsPerPageChange={handleRowsPerPageChange}
+              onSelectAll={customersSelection.handleSelectAll}
+              onSelectOne={customersSelection.handleSelectOne}
+              page={page}
+              rowsPerPage={rowsPerPage}
+              selected={customersSelection.selected}
             />
-          </Grid>
-          <Grid
-            xs={12}
-            sm={6}
-            lg={3}
-          >
-            <OverviewTotalCustomers
-              difference={16}
-              positive={false}
-              sx={{ height: '100%' }}
-              value="1.6k"
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            sm={6}
-            lg={3}
-          >
-            <OverviewTasksProgress
-              sx={{ height: '100%' }}
-              value={75.5}
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            sm={6}
-            lg={3}
-          >
-            <OverviewTotalProfit
-              sx={{ height: '100%' }}
-              value="$15k"
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            lg={8}
-          >
-            <OverviewSales
-              chartSeries={[
-                {
-                  name: 'This year',
-                  data: [18, 16, 5, 8, 3, 14, 14, 16, 17, 19, 18, 20]
-                },
-                {
-                  name: 'Last year',
-                  data: [12, 11, 4, 6, 2, 9, 9, 10, 11, 12, 13, 13]
-                }
-              ]}
-              sx={{ height: '100%' }}
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            md={6}
-            lg={4}
-          >
-            <OverviewTraffic
-              chartSeries={[63, 15, 22]}
-              labels={['Desktop', 'Tablet', 'Phone']}
-              sx={{ height: '100%' }}
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            md={6}
-            lg={4}
-          >
-            <OverviewLatestProducts
-              products={[
-                {
-                  id: '5ece2c077e39da27658aa8a9',
-                  image: '/assets/products/product-1.png',
-                  name: 'Healthcare Erbology',
-                  updatedAt: subHours(now, 6).getTime()
-                },
-                {
-                  id: '5ece2c0d16f70bff2cf86cd8',
-                  image: '/assets/products/product-2.png',
-                  name: 'Makeup Lancome Rouge',
-                  updatedAt: subDays(subHours(now, 8), 2).getTime()
-                },
-                {
-                  id: 'b393ce1b09c1254c3a92c827',
-                  image: '/assets/products/product-5.png',
-                  name: 'Skincare Soja CO',
-                  updatedAt: subDays(subHours(now, 1), 1).getTime()
-                },
-                {
-                  id: 'a6ede15670da63f49f752c89',
-                  image: '/assets/products/product-6.png',
-                  name: 'Makeup Lipstick',
-                  updatedAt: subDays(subHours(now, 3), 3).getTime()
-                },
-                {
-                  id: 'bcad5524fe3a2f8f8620ceda',
-                  image: '/assets/products/product-7.png',
-                  name: 'Healthcare Ritual',
-                  updatedAt: subDays(subHours(now, 5), 6).getTime()
-                }
-              ]}
-              sx={{ height: '100%' }}
-            />
-          </Grid>
-          <Grid
-            xs={12}
-            md={12}
-            lg={8}
-          >
-            <OverviewLatestOrders
-              orders={[
-                {
-                  id: 'f69f88012978187a6c12897f',
-                  ref: 'DEV1049',
-                  amount: 30.5,
-                  customer: {
-                    name: 'Ekaterina Tankova'
-                  },
-                  createdAt: 1555016400000,
-                  status: 'pending'
-                },
-                {
-                  id: '9eaa1c7dd4433f413c308ce2',
-                  ref: 'DEV1048',
-                  amount: 25.1,
-                  customer: {
-                    name: 'Cao Yu'
-                  },
-                  createdAt: 1555016400000,
-                  status: 'delivered'
-                },
-                {
-                  id: '01a5230c811bd04996ce7c13',
-                  ref: 'DEV1047',
-                  amount: 10.99,
-                  customer: {
-                    name: 'Alexa Richardson'
-                  },
-                  createdAt: 1554930000000,
-                  status: 'refunded'
-                },
-                {
-                  id: '1f4e1bd0a87cea23cdb83d18',
-                  ref: 'DEV1046',
-                  amount: 96.43,
-                  customer: {
-                    name: 'Anje Keizer'
-                  },
-                  createdAt: 1554757200000,
-                  status: 'pending'
-                },
-                {
-                  id: '9f974f239d29ede969367103',
-                  ref: 'DEV1045',
-                  amount: 32.54,
-                  customer: {
-                    name: 'Clarke Gillebert'
-                  },
-                  createdAt: 1554670800000,
-                  status: 'delivered'
-                },
-                {
-                  id: 'ffc83c1560ec2f66a1c05596',
-                  ref: 'DEV1044',
-                  amount: 16.76,
-                  customer: {
-                    name: 'Adam Denisov'
-                  },
-                  createdAt: 1554670800000,
-                  status: 'delivered'
-                }
-              ]}
-              sx={{ height: '100%' }}
-            />
-          </Grid>
-        </Grid>
-      </Container>
-    </Box>
-  </>
-);
+          </Stack>
+        </Container>
+      </Box>
+    </>
+  );
+};
 
 Page.getLayout = (page) => (
   <DashboardLayout>
