--- src/sections/overview/overview-total-profit.js
@@ -1,48 +1,0 @@
-import PropTypes from 'prop-types';
-import CurrencyDollarIcon from '@heroicons/react/24/solid/CurrencyDollarIcon';
-import { Avatar, Card, CardContent, Stack, SvgIcon, Typography } from '@mui/material';
-
-export const OverviewTotalProfit = (props) => {
-  const { value, sx } = props;
-
-  return (
-    <Card sx={sx}>
-      <CardContent>
-        <Stack
-          alignItems="flex-start"
-          direction="row"
-          justifyContent="space-between"
-          spacing={3}
-        >
-          <Stack spacing={1}>
-            <Typography
-              color="text.secondary"
-              variant="overline"
-            >
-              Total Profit
-            </Typography>
-            <Typography variant="h4">
-              {value}
-            </Typography>
-          </Stack>
-          <Avatar
-            sx={{
-              backgroundColor: 'primary.main',
-              height: 56,
-              width: 56
-            }}
-          >
-            <SvgIcon>
-              <CurrencyDollarIcon />
-            </SvgIcon>
-          </Avatar>
-        </Stack>
-      </CardContent>
-    </Card>
-  );
-};
-
-OverviewTotalProfit.propTypes = {
-  value: PropTypes.string,
-  sx: PropTypes.object
-};
